# packages/openapi/openapi.yaml
openapi: 3.1.0
...
info:
  title: UBAG REST API
  version: 0.1.0
  summary: Initial UBAG v1 REST contract.
  description: |
    Contract-first REST surface for UBAG v0/v1 workers. This file defines
    route stability, request and response envelopes, stable errors, and job
    event retrieval for the gateway, SDK, worker, CLI, and dashboard surfaces.
  license:
    name: Apache-2.0 for public contract artifacts
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.ubag.dev
    description: Public production API placeholder.
...
security:
  - bearerAuth: []
paths:
...
  /v1/workflows:
    get:
      tags: [Workflows]
      operationId: listWorkflows
      summary: List workflow definitions.
      parameters:
        - $ref: "#/components/parameters/UbagApiVersion"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Workflow records returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/cache:
    get:
      tags: [Cache]
      operationId: getCacheStatus
      summary: Get cache status and visible entries.
      parameters:
        - $ref: "#/components/parameters/UbagApiVersion"
      responses:
        "200":
          description: Cache status returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CacheStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: App-secret or token bearer credential issued by UBAG.
  parameters:
    UbagApiVersion:
      name: Ubag-Api-Version
      in: header
      required: false
      schema:
        type: string
        pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
        examples: ["2026-05-22"]
      description: Requested behavior version. Request bodies with api_version must match this header.
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 512
      description: Opaque pagination cursor.
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
      description: Maximum number of records to return.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid caller credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    MethodNotAllowed:
      description: The route exists but does not allow this HTTP method.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Caller exceeded an applicable rate limit.
      headers:
        Retry-After:
          description: Seconds until the caller should retry.
          schema:
            type: integer
            minimum: 0
        RateLimit-Limit:
          description: Maximum requests in the current window.
          schema:
            type: integer
            minimum: 0
        RateLimit-Remaining:
          description: Requests remaining in the current window.
          schema:
            type: integer
            minimum: 0
        RateLimit-Reset:
          description: Epoch seconds when the current rate-limit window resets.
          schema:
            type: integer
            minimum: 0
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Unexpected service error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      $ref: ../shared-schemas/schemas/error.schema.json
...
    CollectionResponse:
      type: object
      additionalProperties: false
      required: [api_version, kind, data, next_cursor, trace_id]
      properties:
        api_version:
          type: string
          pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
        kind:
          type: string
          enum:
            - events
            - workflows
            - templates
            - targets
            - adapters
            - apps
            - devices
            - webhooks
            - audit
        data:
          type: array
          items:
            type: object
            additionalProperties: true
        next_cursor:
          type: ["string", "null"]
        trace_id:
          type: string
          minLength: 1
    CacheStatus:
      type: object
      additionalProperties: false
      required: [api_version, profile, enabled, entries, trace_id]
      properties:
        api_version:
          type: string
          pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
        profile:
          type: string
          enum: [edge, small, standard, enterprise]
        enabled:
          type: boolean
        entries:
          type: array
          items:
            type: object
            additionalProperties: true
        trace_id:
          type: string
          minLength: 1
...